+++ CHANGES.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+3.4.6 (unreleased)
+------------------
+
+New:
+
+- *add item here*
+
+Fixes:
+
+- *add item here*
+
+
 3.4.5 (2015-11-05)
 ------------------